renames to command palette and docs
