+++ tests/command_palette/test_declare_sources.py
@@ -1,24 +1,19 @@
 from textual.app import App
-from textual.command_palette import (
-    CommandMatches,
-    CommandPalette,
-    CommandSource,
-    CommandSourceHit,
-)
+from textual.command import CommandPalette, Hit, Hits, Source
 from textual.screen import Screen
 
 
 async def test_sources_with_no_known_screen() -> None:
     """A command palette with no known screen should have an empty source set."""
-    assert CommandPalette()._sources == set()
+    assert CommandPalette()._source_classes == set()
 
 
-class ExampleCommandSource(CommandSource):
-    async def search(self, _: str) -> CommandMatches:
+class ExampleCommandSource(Source):
+    async def search(self, _: str) -> Hits:
         def goes_nowhere_does_nothing() -> None:
             pass
 
-        yield CommandSourceHit(1, "Hit", goes_nowhere_does_nothing, "Hit")
+        yield Hit(1, "Hit", goes_nowhere_does_nothing, "Hit")
 
 
 class AppWithActiveCommandPalette(App[None]):
@@ -33,7 +28,9 @@ class AppWithNoSources(AppWithActiveCommandPalette):
 async def test_no_app_command_sources() -> None:
     """An app with no sources declared should work fine."""
     async with AppWithNoSources().run_test() as pilot:
-        assert pilot.app.query_one(CommandPalette)._sources == App.COMMAND_SOURCES
+        assert (
+            pilot.app.query_one(CommandPalette)._source_classes == App.COMMAND_SOURCES
+        )
 
 
 class AppWithSources(AppWithActiveCommandPalette):
@@ -44,7 +41,7 @@ async def test_app_command_sources() -> None:
     """Command sources declared on an app should be in the command palette."""
     async with AppWithSources().run_test() as pilot:
         assert (
-            pilot.app.query_one(CommandPalette)._sources
+            pilot.app.query_one(CommandPalette)._source_classes
             == AppWithSources.COMMAND_SOURCES
         )
 
@@ -66,7 +63,9 @@ class ScreenWithNoSources(Screen[None]):
 async def test_no_screen_command_sources() -> None:
     """An app with a screen with no sources declared should work fine."""
     async with AppWithInitialScreen(ScreenWithNoSources()).run_test() as pilot:
-        assert pilot.app.query_one(CommandPalette)._sources == App.COMMAND_SOURCES
+        assert (
+            pilot.app.query_one(CommandPalette)._source_classes == App.COMMAND_SOURCES
+        )
 
 
 class ScreenWithSources(ScreenWithNoSources):
@@ -77,7 +76,7 @@ async def test_screen_command_sources() -> None:
     """Command sources declared on a screen should be in the command palette."""
     async with AppWithInitialScreen(ScreenWithSources()).run_test() as pilot:
         assert (
-            pilot.app.query_one(CommandPalette)._sources
+            pilot.app.query_one(CommandPalette)._source_classes
             == App.COMMAND_SOURCES | ScreenWithSources.COMMAND_SOURCES
         )
 
@@ -97,6 +96,6 @@ async def test_app_and_screen_command_sources_combine() -> None:
     """If an app and the screen have command sources they should combine."""
     async with CombinedSourceApp().run_test() as pilot:
         assert (
-            pilot.app.query_one(CommandPalette)._sources
+            pilot.app.query_one(CommandPalette)._source_classes
             == CombinedSourceApp.COMMAND_SOURCES | ScreenWithSources.COMMAND_SOURCES
         )
