+++ tests/command_palette/test_escaping.py
@@ -1,18 +1,13 @@
 from textual.app import App
-from textual.command_palette import (
-    CommandMatches,
-    CommandPalette,
-    CommandSource,
-    CommandSourceHit,
-)
+from textual.command import CommandPalette, Hit, Hits, Source
 
 
-class SimpleSource(CommandSource):
-    async def search(self, query: str) -> CommandMatches:
+class SimpleSource(Source):
+    async def search(self, query: str) -> Hits:
         def goes_nowhere_does_nothing() -> None:
             pass
 
-        yield CommandSourceHit(1, query, goes_nowhere_does_nothing, query)
+        yield Hit(1, query, goes_nowhere_does_nothing, query)
 
 
 class CommandPaletteApp(App[None]):
