+++ tests/command_palette/test_interaction.py
@@ -1,20 +1,14 @@
 from textual.app import App
-from textual.command_palette import (
-    CommandList,
-    CommandMatches,
-    CommandPalette,
-    CommandSource,
-    CommandSourceHit,
-)
+from textual.command import CommandList, CommandPalette, Hit, Hits, Source
 
 
-class SimpleSource(CommandSource):
-    async def search(self, query: str) -> CommandMatches:
+class SimpleSource(Source):
+    async def search(self, query: str) -> Hits:
         def goes_nowhere_does_nothing() -> None:
             pass
 
         for _ in range(100):
-            yield CommandSourceHit(1, query, goes_nowhere_does_nothing, query)
+            yield Hit(1, query, goes_nowhere_does_nothing, query)
 
 
 class CommandPaletteApp(App[None]):
