+++ tests/snapshot_tests/snapshot_apps/command_palette.py
@@ -1,25 +1,29 @@
 from textual.app import App
-from textual.command_palette import CommandSource, CommandMatches, CommandSourceHit
+from textual.command import Source, Hits, Hit
 
-class TestSource(CommandSource):
 
+class TestSource(Source):
     def goes_nowhere_does_nothing(self) -> None:
         pass
 
-    async def search(self, query: str) -> CommandMatches:
+    async def search(self, query: str) -> Hits:
         matcher = self.matcher(query)
         for n in range(10):
             command = f"This is a test of this code {n}"
-            yield CommandSourceHit(
-                n/10, matcher.highlight(command), self.goes_nowhere_does_nothing, command
+            yield Hit(
+                n / 10,
+                matcher.highlight(command),
+                self.goes_nowhere_does_nothing,
+                command,
             )
 
-class CommandPaletteApp(App[None]):
 
+class CommandPaletteApp(App[None]):
     COMMAND_SOURCES = {TestSource}
 
     def on_mount(self) -> None:
         self.action_command_palette()
+
 
 if __name__ == "__main__":
     CommandPaletteApp().run()
