+++ tests/snapshot_tests/test_snapshots.py
@@ -600,14 +600,15 @@ def test_tooltips_in_compound_widgets(snap_compare):
 
 
 def test_command_palette(snap_compare) -> None:
-
-    from textual.command_palette import CommandPalette
+    from textual.command import CommandPalette
 
     async def run_before(pilot) -> None:
         await pilot.press("ctrl+@")
         await pilot.press("A")
         await pilot.app.query_one(CommandPalette).workers.wait_for_complete()
+
     assert snap_compare(SNAPSHOT_APPS_DIR / "command_palette.py", run_before=run_before)
+
 
 # --- textual-dev library preview tests ---
